requirements.txt
absl-py==0.13.0
astunparse==1.6.3
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
clang==5.0
cycler==0.10.0
Cython==0.29.24
deepctr==0.9.0
flatbuffers==1.12
gast==0.3.3
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
joblib==1.0.1
keras==2.6.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lxml==4.6.3
Markdown==3.3.4
mathplot==0.1
matplotlib==3.4.2
numpy==1.21.2
nyoka==5.0.0
oauthlib==3.1.1
opt-einsum==3.3.0
pandas==1.3.0
Pillow==8.3.1
protobuf==3.17.3
py4j==0.10.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pypmml==0.9.11
python-dateutil==2.8.1
python-gflags==3.1.2
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.4.1
seaborn==0.11.1
six==1.15.0
sklearn-pandas==2.2.0
sklearn2pmml==0.73.5
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.4.0
termcolor==1.1.0
threadpoolctl==2.1.0
toad==0.0.65
typing-extensions==3.7.4.3
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.12.1
xgboost==1.3.3